requirements.txt
streamlit==1.32.0
pandas==2.2.1
pdfplumber==0.11.0
langchain-community==0.0.27
openpyxl==3.1.2
requests==2.31.0
mistralai==0.1.8
python-dotenv==1.0.1
